Move language modeling to a separate service
This allows other parts of Lingetic not written
in Java to use these.
