+++ lingetic-spring-backend/config/application.properties
@@ -1,6 +1,8 @@
 spring.application.name=lingetic
 server.port=8000
 
+app.environment=${ENVIRONMENT}
+
 spring.web.cors.allowed-origins=${FRONTEND_URL}
 spring.main.allow-circular-references=true
