+++ lingetic-spring-backend/src/main/java/com/munetmo/lingetic/LanguageService/Entities/LanguageModels/EnglishLanguageModel.java
@@ -1,6 +1,6 @@
-package com.munetmo.lingetic.LanguageTestService.Entities.LanguageModels;
+package com.munetmo.lingetic.LanguageService.Entities.LanguageModels;
 
-import com.munetmo.lingetic.LanguageTestService.Entities.Language;
+import com.munetmo.lingetic.LanguageService.Entities.Language;
 
 import java.util.Arrays;
 import java.util.Locale;
