+++ lingetic-spring-backend/src/main/java/com/munetmo/lingetic/LanguageService/Entities/Token.java
@@ -1,0 +1,4 @@
+package com.munetmo.lingetic.LanguageService.Entities;
+
+public class Token {
+}
