+++ lingetic-spring-backend/src/main/java/com/munetmo/lingetic/LanguageTestService/UseCases/GetQuestionListsForLanguageUseCase.java
@@ -1,6 +1,6 @@
 package com.munetmo.lingetic.LanguageTestService.UseCases;
 
-import com.munetmo.lingetic.LanguageTestService.Entities.Language;
+import com.munetmo.lingetic.LanguageService.Entities.Language;
 import com.munetmo.lingetic.LanguageTestService.Entities.QuestionList;
 import com.munetmo.lingetic.LanguageTestService.Repositories.QuestionListRepository;
