+++ lingetic-spring-backend/src/main/java/com/munetmo/lingetic/infra/auth/SecurityConfig.java
@@ -38,6 +38,7 @@ public class SecurityConfig {
                 .csrf(AbstractHttpConfigurer::disable)
                 .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .authorizeHttpRequests(auth -> auth
+                        .requestMatchers("/language-service/**").permitAll()
                         .requestMatchers("/health-service/wakeup").permitAll()
                         .requestMatchers("/language-test-service/questions/review").permitAll()
                         .anyRequest().authenticated())
