+++ lingetic-spring-backend/src/test/java/com/munetmo/lingetic/LanguageTestService/Entities/QuestionReviewTest.java
@@ -6,6 +6,7 @@ import java.time.Duration;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 
+import com.munetmo.lingetic.LanguageService.Entities.Language;
 import org.junit.jupiter.api.Test;
 
 class QuestionReviewTest {
